pyproject.toml
[tool.poetry]
name = "brief-scraping-scrapy"
version = "0.1.0"
description = ""
authors = ["DaiTensa <dai.tensaout@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Scrapy = "^2.11.2"
mysql-connector-python = "^8.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
